Chilling units

Compute chilling units accumulated by a plant, according to a beta-shaped function.

chillingUnits(temp, temp.min, temp.max, mu, s)

Arguments

temp
the vector of temperatures to accumulate

temp.min
the threshold above which the plant is accumulating chilling units

temp.max
the threshold below which the plant is accumulating chilling units

mu
the temperature at which the plant accumulates the highest number of chilling units (the mode of the underlying Beta distribution)

s
the sample size of the underlying Beta distribution

Value

a vector with the chilling units corresponding to each input temperature

Details

The accumulation of chilling units is expressed as the Beta law probability distribution function (pdf). The underlying Beta law is re-parametrized using the mode and the so-called sample size. More precisely, using the following definition for the pdf of the Beta law with parameters a and b: $$\frac{\Gamma(a+b)}{\Gamma (a)+\Gamma (b)} \ x^{a-1}(1-x)^{1-b}$$ for \(0 \leq x \leq 1\), \(a>0\) and \(b >0\). Now, we further assume here that \(a > 1\) and \(b > 1\), so that the mode exists, and we parametrized the Beta distribution using \(\mu\) and \(s\) defined as: $$\mu = \frac{a-1}{a+b-2}, \quad s = a+b$$

Since the Beta law has its support on \([0,1]\), it is evaluated at (x-temp.min)/(temp.max-temp.min), where x is the recorded temperature